api/src/requirements.txt
firebase_admin==5.2.0
functions_framework==3.0.0
ImageHash==4.2.1
Pillow==9.1.0
requests==2.27.1
